Fix double program release in base model (#1223)

+++ modules/core/src/lib/base-model.js
@@ -306,12 +306,8 @@ export default class BaseModel {
       return;
     }
 
-    if (this.program) {
-      if (this.programManager) {
-        this.programManager.release(this.program);
-      } else {
-        this.program.delete();
-      }
+    if (this.program && !this.programManager) {
+      this.program.delete();
     }
 
     this.program = program;
